requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
markdown>=3.5.1
python-multipart>=0.0.6
rich>=13.5.2  # CLI formatting
typing-extensions>=4.8.0
beautifulsoup4>=4.14.3


# Formatting
flake8>=6.1.0
black>=23.11.0
mypy>=1.7.1
pre-commit>=3.5.0